% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Change.Point}
\alias{Change.Point}
\title{Change.Point}
\usage{
Change.Point(x, y)
}
\arguments{
\item{x}{: x axis vector}

\item{x}{: y axis vector}
}
\value{
change point index in x vector,
Note that the first and the last point will be always change point
}
\description{
Change.Point
}
