% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Convert.Bin.To.Domain}
\alias{Convert.Bin.To.Domain}
\title{Convert.Bin.To.Domain}
\usage{
Convert.Bin.To.Domain(bins, signal.idx, gap.idx, pvalues = NULL,
  pvalue.cut = NULL)
}
\arguments{
\item{bins}{: bin information}

\item{signal.idx}{: signal index}

\item{pvalues}{: pvalue vector}

\item{pvalue.cut}{: pvalue threshold}

\item{signal.idx}{: gap index}
}
\value{
dataframe storing domain information
}
\description{
Convert.Bin.To.Domain
}
