% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_mat_from_b.R
\name{correct_mat_from_b}
\alias{correct_mat_from_b}
\title{Apply a bias correction to a HiC contact matrix}
\usage{
correct_mat_from_b(mat, b)
}
\arguments{
\item{mat}{HiC contact map matrix}

\item{b}{Vecor of biases}
}
\value{
A corrected HiC contact matrix
}
\description{
This function takes a contact matrix and a vector of biases and returns the corrected contact matrix
}
\details{
Given a raw contact matrix \eqn{R_{ij}} and a vector of biases \eqn{b_i}, the corrected matrix is \eqn{C_{ij} = R_{ij} / (b_i b_j)}
}
\examples{
plot(0)
}
