% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Detect.Local.Extreme}
\alias{Detect.Local.Extreme}
\title{Detect.Local.Extreme}
\usage{
Detect.Local.Extreme(x)
}
\arguments{
\item{x}{: original signal to find local minima}
}
\value{
vector of local extrme, -1 if the index is local minimum, 1 if the index is local maxima, 0 otherwise.
}
\description{
Detect.Local.Extreme
}
