% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evec_dist.R
\name{evec_dist}
\alias{evec_dist}
\title{Compute the distance between two eigenvectors.}
\usage{
evec_dist(ev1, ev2)
}
\arguments{
\item{ev1}{A eigen vector.}

\item{ev2}{Another eigen vector.}
}
\value{
The euclidean distance between the vectors..
}
\description{
This function computes the euclidian distance between two eigenvectros, taking into account their arbitrary relative signs.
}
\examples{
plot(0)
}
