% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_max.R
\name{first_max}
\alias{first_max}
\title{Find first local minimum}
\usage{
first_max(x, plot = FALSE, zero_rm = FALSE)
}
\arguments{
\item{x}{A numeric \code{vector} of values.}

\item{plot}{\code{logical} indicating if histogram should be plotted.}

\item{zero_rm}{\code{logical} indicating if zeros should be removed from calculation.}
}
\value{
A numeric value of the lower local maximum value of the density estimation.
}
\description{
This funciton computes the lowest maximum value of the density estimate of the input numeric vector
}
\examples{
plot(0)
}
