% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitcnv.R
\name{fitcnv}
\alias{fitcnv}
\title{Estimate copy number}
\usage{
fitcnv(x)
}
\arguments{
\item{x}{A vector of total contacts per bin. It should be ordered (at least inside each chromosome).}
}
\value{
A \code{list} with two elements. The first one refers to the average number of counts corresponding to a copy number of two. The second one is a vector of the same length as \code{x} with the estimated copy number.
}
\description{
This function takes a vector with the total number of contacts per bin and estimates the copy number
}
\examples{
plot(0)
}
