% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Get.Diamond.Matrix}
\alias{Get.Diamond.Matrix}
\title{Get.Diamond.Matrix}
\usage{
Get.Diamond.Matrix(mat.data, i, size)
}
\arguments{
\item{mat.data}{: N by N matrix, where each element indicate contact frequency}

\item{i}{:integer, bin index}

\item{size}{: integer, window size to expand from bin}
}
\value{
: matrix.
}
\description{
Get.Diamond.Matrix
}
