% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Get.Diamond.Matrix2}
\alias{Get.Diamond.Matrix2}
\title{Get.Diamond.Matrix2}
\usage{
Get.Diamond.Matrix2(mat.data, i, size)
}
\arguments{
\item{mat.data}{: matrix data}

\item{i}{: bin index}

\item{size}{: size of window to extend}
}
\value{
diamond matrix
}
\description{
Get.Diamond.Matrix2
}
