% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Get.Downstream.Triangle}
\alias{Get.Downstream.Triangle}
\title{Get.Downstream.Triangle}
\usage{
Get.Downstream.Triangle(mat.data, i, size)
}
\arguments{
\item{mat.data}{: matrix data}

\item{i}{: bin index}

\item{size}{: size of window to extend}
}
\value{
downstream triangle matrix
}
\description{
Get.Downstream.Triangle
}
