% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Get.Pvalue}
\alias{Get.Pvalue}
\title{Get.Pvalue}
\usage{
Get.Pvalue(matrix.data, size, scale = 1)
}
\arguments{
\item{matrix.data}{: matrix}

\item{size}{: size to extend}

\item{scale}{: scale parameter if necessary. deprecated parameter}
}
\value{
computed p-value vector
}
\description{
Get.Pvalue
}
