% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts_matrix.R
\name{get_contacts_matrix}
\alias{get_contacts_matrix}
\title{Make a contact matrix from HiC data}
\usage{
get_contacts_matrix(inbam, resolution, pos, region = NULL, whole = F,
  filtin = 0, filtex = 783)
}
\arguments{
\item{inbam}{HiC-BAM file}

\item{resolution}{Desired resolution (bin size)}

\item{pos}{A vector with the genomic position IDs to be included in the output. It can be the \code{bin} column of the output from \code{\link{make_bins}}}

\item{region}{Region of interest following the format \code{chr|chr:start-end}. If \code{NULL}, take the entire genome}

\item{whole}{Logical indicating if the entire available genome should be retrieved (rectangular contact map) or only the region of interest indicated with \code{region} (square contact map). It should be consistent the \code{pos} argument}

\item{filtin}{Integer value indicating the required bits for the HiC-BAM FLAG (include the corresponding contacts in the output)}

\item{filtex}{Integer value indicating the excluded bits for the HiC-BAM FLAG (exclude the corresponding contacts in the output)}
}
\value{
A named \code{sparse_matrix} containing the number of contacts per pair of genomic bins at the requested region
}
\description{
This function takes a HiC-BAM file and creates a contact map as a \code{sparse_ matrix} with \code{dimnames} corresponding to the genomic bins
}
\examples{
plot(0)
}

