% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contacts_per_bin.R
\name{get_contacts_per_bin}
\alias{get_contacts_per_bin}
\title{Retrieve HiC data per genomic bin}
\usage{
get_contacts_per_bin(
  inbam,
  resolution,
  region = NULL,
  pos,
  filtin = 0,
  filtex = 783
)
}
\arguments{
\item{inbam}{HiC-BAM file}

\item{resolution}{Desired resolution (bin size)}

\item{region}{Region of interest following the format \code{chr|chr:start-end}. If \code{NULL}, take the entire genome}

\item{pos}{A vector with the genomic position IDs to be included in the output. It can be the \code{bin} column of the output from \code{\link{make_bins}}}

\item{filtin}{Integer value indicating the required bits for the HiC-BAM FLAG (include the corresponding contacts in the output)}

\item{filtex}{Integer value indicating the excluded bits for the HiC-BAM FLAG (exclude the corresponding contacts in the output)}
}
\value{
A named \code{vector} containing the total number of contacts per genomic bin at the requested region
}
\description{
This function takes a HiC-BAM file and creates a nammed \code{vector} with the total number of contact per genomic bin
}
\examples{
plot(0)
}
