% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_decay.R
\name{get_distance_decay}
\alias{get_distance_decay}
\title{Compute distance decay of a contact matrix}
\usage{
get_distance_decay(mat, reso)
}
\arguments{
\item{mat}{A contact matrix (could be the output of \code{\link{get_contacts_matrix}})}

\item{reso}{Bin size (resolution) in bp}
}
\value{
If \code{mat} contains more than one chromosome, a list of \code{data.frame}s. If not, a \code{data.frame} with the distance \code{d} (in bp) and the expected number of counts \code{e}.
}
\description{
This function takes a contact matrix (tipically the output of \code{\link{get_contacts_matrix}} and returns the distance decay
}
\examples{
plot(0)
}
