% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expected_matrix.R
\name{get_expected_matrix}
\alias{get_expected_matrix}
\title{Compute expected matrix (by distance decay).}
\usage{
get_expected_matrix(mat, dd)
}
\arguments{
\item{mat}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}. Currently only one chromosome.}

\item{dd}{A \code{data.frame} with two columns: \code{d} (distance in bp) and \code{e} (expected number of contacts.}
}
\value{
A \code{dgTMatrix} corresponding to expected counts.
}
\description{
This function takes a sparse matrix of contacts and data frame with the expected counts per distance and returns a expected matrix.
}
\examples{
plot(0)
}
