% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kr_sparse.R
\name{kr_sparse}
\alias{kr_sparse}
\title{Perform the Knight-Ruiz normalization.}
\usage{
kr_sparse(A)
}
\arguments{
\item{A}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}
}
\value{
A A \code{dgTMatrix} of normalized contacts.
}
\description{
This function takes a contact matrix and returns its KR normalized version.
}
\details{
This is a re-implementation of \url{https://github.com/dozmorovlab/HiCcompare/blob/master/R/KRnormalization.R} using sparse matrices.
}
\examples{
plot(0)
}
\references{
\url{https://doi.org/10.1093/imanum/drs019}
}
