% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian_matrix.R
\name{laplacian_matrix}
\alias{laplacian_matrix}
\title{Get the associated laplacian matrix.}
\usage{
laplacian_matrix(x)
}
\arguments{
\item{x}{A \code{dgTMatrix} of contacts.}
}
\value{
The associated \code{dgTMatrix} laplacian matrix.
}
\description{
This function takes sparse matrix of contacts and returns its corresponding Laplacian matrix.
}
\details{
This is an auxiliary function used by \code{reproducibility_score}.
}
\examples{
plot(0)
}
