% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfinite.R
\name{logfinite}
\alias{logfinite}
\title{Log-transform a vector and substitute infinite values by \code{NA}}
\usage{
logfinite(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector with the log10 transformed values where the infinite ones have been replaced by \code{NA}
}
\description{
This function takes a vector of values and returns the log10 transformed values replacing infinite values by \code{NA}
}
\examples{
x <- c(10, 1000, 4, 0, 0.1)
logfinite(x)
}

