% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logpseudo.R
\name{logpseudo}
\alias{logpseudo}
\title{Add pseudo counts to a vector and Log-transform it}
\usage{
logpseudo(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
A vector with the log10 transformation of the original values plus a pseudo count equal to half of the minimum of the positive original values
}
\description{
This function takes a vector of values and returns the log10 transformed values adding a pseudo count to the values prior to the trasformation
}
\details{
The pseudo count equals half the minimum value, exluding zeros.
}
\examples{
x <- c(10, 1000, 4, 0, 0.1)
logpseudo(x)
}
