% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{Bin the genome}
\usage{
make_bins(inbam, resolution)
}
\arguments{
\item{inbam}{HiC-BAM file (only the header will be used)}

\item{resolution}{Desired resolution (bin size)}
}
\value{
A \code{data.frame} containing chromosome, position and bin ID
}
\description{
This function takes a HiC-BAM file and creates a data.frame with the genome binned at the desired resolution
}
\examples{
plot(0)
}

