% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2df.R
\name{mat2df}
\alias{mat2df}
\title{Transform the contacts of a \code{dgTMatrix} into a \code{data.frame}}
\usage{
mat2df(mat, b = NULL)
}
\arguments{
\item{mat}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{b}{An optional \code{vector} of genomic position (otherwhise, it uses the infromation contained in the \code{row.names}.}
}
\value{
A \code{data.frame} with the two coordinates (\code{i} and \code{j}) and the number of contacts (\code{x}).
}
\description{
This function takes a sparse matrix of contacts and transforms it into a \code{data.frame}. It works only with matrices of one chromosome.
}
\details{
Please note that mat is expected to be a matrix of one chromosome (you could obtain one using \code{subset_matrix}).
}
\examples{
plot(0)
}
