% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_res_similarity.R
\name{matrix_res_similarity}
\alias{matrix_res_similarity}
\title{Compute the similarity between raw contact matrices of different resolutions for the same sample}
\usage{
matrix_res_similarity(mtx1, mtx2, res1, res2, method = "F")
}
\arguments{
\item{mtx1}{Raw contact matrix 1 (dense or sparse)}

\item{mtx2}{Raw contact matrix 2 (dense or sparse)}

\item{res1}{Resolution (i.e. bin size) of mtx1}

\item{res2}{Resolution (i.e. bin size) of mtx2}

\item{method}{Matrix norm method. Default: Frobenius}
}
\description{
Compute the similarity between raw contact matrices of different resolutions for the same sample
}
