% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_matrices.R
\name{paste_matrices}
\alias{paste_matrices}
\title{Create a \code{data.frame} with the contacts of two matrices (including zeros).}
\usage{
paste_matrices(smat1, smat2, mm)
}
\arguments{
\item{smat1}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{smat2}{Another \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{mm}{The maximum distance (in bins) to truncate the data.}
}
\value{
A \code{data.frame} with the two coordinates (\code{i} and \code{j}) and the number of contacts of each matrix (\code{x1} and \code{x2}).
}
\description{
This function takes two sparse matrix of contacts and a maximum distance (in bins) and returns a \code{data.frame} with the contacts of the tow matrices for each pair of positions.
}
\details{
This is an auxiliary function used by \code{scc}.
}
\examples{
plot(0)
}
