% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot HiC contact matrix}
\usage{
plot_matrix(
  mat,
  coord,
  resolution,
  transformation = logfinite,
  color = colorRampPalette(c("white", "red"))(100),
  sym = FALSE,
  trim = 0.01,
  rotate = FALSE,
  unit_x_axis = 1e+06,
  label_x_axis = "Genomic Position / Mbp",
  na.col = "white",
  ...
)
}
\arguments{
\item{mat}{HiC contact matrix (it could be the output of \code{\link{get_contacts_matrix}})}

\item{coord}{A vector of size two with the start and end coordinates of the desired region to plot}

\item{resolution}{Resolution (bin size) in bp}

\item{transformation}{Transformation to apply to the contacts prior to the heatmap representation}

\item{color}{Vector of colors}

\item{sym}{Logical indicating if the color scale should be symmetrical (around 0)}

\item{trim}{Numeric proportion of the data that should be 'flattened' on both extremes}

\item{rotate}{Logical indicating if the matrix should be rotated}

\item{unit_x_axis}{Numeric unit for the X axis}

\item{label_x_axis}{Character X axis label}

\item{na.col}{Color to depict \code{NA} values}

\item{...}{Further arguments to be passed to \code{\link{plot}}}
}
\description{
This function takes a HiC contact matrix and creates heatmap-like prepresentation of it
}
\examples{
plot(0)
}
