% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix_region.R
\name{plot_matrix_region}
\alias{plot_matrix_region}
\title{Plot a desired region of a contact matrix using genomic coordinates}
\usage{
plot_matrix_region(mat, bins, coord, chrom = strsplit(coord, "[:-]", perl
  = TRUE)[[1]][1], start = as.integer(strsplit(coord, "[:-]", perl =
  TRUE)[[1]][2]), end = as.integer(strsplit(coord, "[:-]", perl =
  TRUE)[[1]][3]), ...)
}
\arguments{
\item{mat}{HiC contact matrix (it could be the output of \code{\link{get_contacts_matrix}})}

\item{bins}{A \code{data.frame} containing chromosome, position and bin ID}

\item{coord}{A string of the form \code{chr:start-end} specifying the genomic region to plot}

\item{chrom}{Chromosome to plot. Priority over \code{coord},}

\item{start}{Starting coordinate. Priority over \code{coord},}

\item{end}{Ending coordinate. Priority over \code{coord},}

\item{...}{Parameters passed onto \code{plot_matrix} function}
}
\value{
returns
}
\description{
Plot a desired region of a contact matrix using genomic coordinates
}
