% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibility_score.R
\name{reproducibility_score}
\alias{reproducibility_score}
\title{Compute the reproducibility score between two contact matrices.}
\usage{
reproducibility_score(mat1, mat2, r = 20, mipr = 5, na.rm = TRUE)
}
\arguments{
\item{mat1}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{mat2}{Another \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{r}{The number of componentes used in the comparison.}

\item{mipr}{Threshold for the j-inverse participation ratio (IPR).}

\item{na.rm}{\code{logical} indicating if bins with NA should be removed.}
}
\value{
The reproducibility score between two matrices.
}
\description{
This function takes a couple of contact matrices and computes the reproducibility score.
}
\details{
Please note that mat1 and mat2 are expected to be a matrices of one chromosome (you could obtain them using \code{subset_matrix}).
}
\examples{
plot(0)
}
\references{
\url{https://doi.org/10.1093/bioinformatics/btx152}
}
