% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc.R
\name{scc}
\alias{scc}
\title{Compute the stratum adjusted correlation coefficient (SCC) between two contact matrices.}
\usage{
scc(mat1, mat2, mm, resol)
}
\arguments{
\item{mat1}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{mat2}{Another \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{mm}{The maximum distance (in bp) to truncate the data.}

\item{resol}{The matrix resolution (bin size).}
}
\value{
A list with the estimated SCC and associated std.
}
\description{
This function takes a couple of contact matrices and computes the reproducibility score.
}
\details{
Please note that mat1 and mat2 are expected to be a matrices of one chromosome (you could obtain them using \code{subset_matrix}).
}
\examples{
plot(0)
}
\references{
\url{https://doi.org/10.1101/101386 }
}
