% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_matrix.R
\name{subsample_matrix}
\alias{subsample_matrix}
\title{Subsample a HiC contact matrix}
\usage{
subsample_matrix(mat, p = 0.5)
}
\arguments{
\item{mat}{HiC contact map matrix}

\item{p}{Numeric probability of being sub-sampled}
}
\value{
A sub-sampled HiC contact matrix
}
\description{
This function takes a HiC contact matrix and returns a subsample of it. The input matrix should be triangular.
}
\examples{
plot(0)
}

