% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_matrix.R
\name{subset_matrix}
\alias{subset_matrix}
\title{Extract the contacts associated to a region form a acontact matrix.}
\usage{
subset_matrix(mat, region)
}
\arguments{
\item{mat}{A \code{dgTMatrix} of contacts with genomic bins as \code{row.names}.}

\item{region}{A genomic region ("chr" or "chr:start-end").}
}
\value{
A \code{dgTMatrix} corresponding to the desired region.
}
\description{
This function takes a sparse matrix of contacts and a genomic region and extract a sub-matrix with the desired region.
}
\examples{
plot(0)
}
