% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetrize_matrix.R
\name{symmetrize_matrix}
\alias{symmetrize_matrix}
\title{Get the full symmetric matrix from a upper triangular contact matrix}
\usage{
symmetrize_matrix(mat)
}
\arguments{
\item{mat}{HiC contact matrix (should be upper triangular)}
}
\value{
The full symmetric HiC contact matrix corresponding to the input one
}
\description{
This function takes a upper triangular HiC contact matrix and returns the full symmetric matrix.
}
\examples{
plot(0)
}

