% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upper_tri.R
\name{upper_tri}
\alias{upper_tri}
\title{Get upper triangular matrix from a HiC contact matrix}
\usage{
upper_tri(mat)
}
\arguments{
\item{mat}{HiC contact matrix}
}
\value{
The upper triangular matrix of the original one
}
\description{
This function takes a HiC contact matrix and returns the upper triangular matrix. It divides the diagonal counts by 2.
}
\examples{
plot(0)
}

