% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Which.Gap.Region}
\alias{Which.Gap.Region}
\title{Which.Gap.Region}
\usage{
Which.Gap.Region(matrix.data)
}
\arguments{
\item{matrix.data}{: n by n matrix}
}
\value{
gap index
}
\description{
Which.Gap.Region
}
