% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Which.Gap.Region2}
\alias{Which.Gap.Region2}
\title{Which.Gap.Region2}
\usage{
Which.Gap.Region2(matrix.data, w)
}
\arguments{
\item{matrix.data}{: n by n matrix}
}
\value{
gap index
}
\description{
Which.Gap.Region2
}
