% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Which.Gap.Region3}
\alias{Which.Gap.Region3}
\title{Which.Gap.Region3}
\usage{
Which.Gap.Region3(mean.cf)
}
\arguments{
\item{matrix.data}{: n by n matrix}
}
\value{
gap index
}
\description{
Which.Gap.Region3
}
