% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdom.R
\name{Which.process.region}
\alias{Which.process.region}
\title{Which.process.region}
\usage{
Which.process.region(rmv.idx, n_bins, min.size = 3)
}
\arguments{
\item{rmv.idx}{: vector of idx, remove index vector}

\item{n_bins}{: total number of bins}

\item{min.size}{: minimum size of bins}
}
\value{
: data.frame of proc.regions
}
\description{
Which.process.region
}
