% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_min.R
\name{first_min}
\alias{first_min}
\title{Find first local minimum}
\usage{
first_min(x, plot = FALSE)
}
\arguments{
\item{x}{A numeric \code{vector} of values.}

\item{plot}{\code{logical} indicating if histogram should be plotted.}
}
\value{
A numeric value of the lower local minimum value of the density estimation.
}
\description{
This funciton computes the lowest minimum value of the density estimate of theinput numeric vector
}
\examples{
plot(0)
}
