% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oned.R
\name{oned}
\alias{oned}
\title{Compute oned correction}
\usage{
oned(dat, form = tot ~ s(map) + s(cg) + s(res))
}
\arguments{
\item{dat}{A \code{data.frame} with one bin per row containing the total number of contacts and the potential biases as columns.}

\item{form}{A \code{formula} describing the  total number of contacts on LHS and the smoothed biases on the RHS (should be compatible with \code{\link{mgcv::gam}})}
}
\value{
A vector of length \code{nrow(dat)} with the biases to correct for.
}
\description{
This function takes a \code{data.frame} with bin information and returns a vector of  biases to correct for
}
\details{
Please note that the biases returned are the squarerooted so one can directly apply \code{\link{correct_mat_from_b}}
}
\examples{
plot(0)
}

